refactor(dashboard): extract share action handlers into useShareActions hook
